requirements.txt
pip
youtube_dl
PyNaCl
PyGithub
discord.py[voice]
dislash.py